refactor: move openai_request to util file
